implementation of security

+++ tcrs-app/src/main/java/com/tcrs_app/tcrs_app/controllers/AuthenticationController.java
@@ -5,71 +5,43 @@ import com.tcrs_app.tcrs_app.payload.request.RefreshTokenRequest;
 import com.tcrs_app.tcrs_app.payload.request.RegisterRequest;
 import com.tcrs_app.tcrs_app.payload.response.AuthenticationResponse;
 import com.tcrs_app.tcrs_app.payload.response.RefreshTokenResponse;
+import com.tcrs_app.tcrs_app.payload.response.RegisterResponse;
 import com.tcrs_app.tcrs_app.services.AuthenticationService;
 import com.tcrs_app.tcrs_app.services.JwtService;
 import com.tcrs_app.tcrs_app.services.RefreshTokenService;
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.Schema;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.security.SecurityRequirements;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseCookie;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.web.ErrorResponse;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("/api/v1/auth")
-@SecurityRequirements()
 @RequiredArgsConstructor
 public class AuthenticationController {
 
     private final AuthenticationService authenticationService;
     private final RefreshTokenService refreshTokenService;
-    private final AuthenticationManager authenticationManager;
     private final JwtService jwtService;
 
     @PostMapping("/register")
-    public ResponseEntity<AuthenticationResponse> register(@Valid @RequestBody RegisterRequest request) {
-        AuthenticationResponse authenticationResponse = authenticationService.register(request);
-        ResponseCookie jwtCookie = jwtService.generateJwtCookie(authenticationResponse.getAccessToken());
-        ResponseCookie refreshTokenCookie = refreshTokenService.generateRefreshTokenCookie(authenticationResponse.getRefreshToken());
-        return ResponseEntity.ok()
-                .header(HttpHeaders.SET_COOKIE, jwtCookie.toString())
-                .header(HttpHeaders.SET_COOKIE,refreshTokenCookie.toString())
-                .body(authenticationResponse);
+    public ResponseEntity<RegisterResponse> register(@Valid @RequestBody RegisterRequest request) {
+        return ResponseEntity.ok(authenticationService.register(request));
     }
 
     @PostMapping("/authenticate")
-    @Operation(
-            responses = {
-                    @ApiResponse(
-                            description = "Success",
-                            responseCode = "200"
-                    ),
-                    @ApiResponse(
-                            description = "Unauthorized",
-                            responseCode = "401",
-                            content = {@Content(schema = @Schema(implementation = ErrorResponse.class), mediaType = "application/json")}
-                    )
-            }
-    )
     public ResponseEntity<AuthenticationResponse> authenticate(@RequestBody AuthenticationRequest request) {
         AuthenticationResponse authenticationResponse = authenticationService.authenticate(request);
         ResponseCookie jwtCookie = jwtService.generateJwtCookie(authenticationResponse.getAccessToken());
         ResponseCookie refreshTokenCookie = refreshTokenService.generateRefreshTokenCookie(authenticationResponse.getRefreshToken());
         return ResponseEntity.ok()
-                .header(HttpHeaders.SET_COOKIE,jwtCookie.toString())
-                .header(HttpHeaders.SET_COOKIE,refreshTokenCookie.toString())
+                .header(HttpHeaders.SET_COOKIE, jwtCookie.toString())
+                .header(HttpHeaders.SET_COOKIE, refreshTokenCookie.toString())
                 .body(authenticationResponse);
     }
+
     @PostMapping("/refresh-token")
     public ResponseEntity<RefreshTokenResponse> refreshToken(@RequestBody RefreshTokenRequest request) {
         return ResponseEntity.ok(refreshTokenService.generateNewToken(request));
@@ -80,30 +52,23 @@ public class AuthenticationController {
         String refreshToken = refreshTokenService.getRefreshTokenFromCookies(request);
         RefreshTokenResponse refreshTokenResponse = refreshTokenService
                 .generateNewToken(new RefreshTokenRequest(refreshToken));
-        ResponseCookie NewJwtCookie = jwtService.generateJwtCookie(refreshTokenResponse.getAccessToken());
+        ResponseCookie newJwtCookie = jwtService.generateJwtCookie(refreshTokenResponse.getAccessToken());
         return ResponseEntity.ok()
-                .header(HttpHeaders.SET_COOKIE, NewJwtCookie.toString())
+                .header(HttpHeaders.SET_COOKIE, newJwtCookie.toString())
                 .build();
-    }
-    @GetMapping("/info")
-    public Authentication getAuthentication(@RequestBody AuthenticationRequest request){
-        return     authenticationManager.authenticate(
-                new UsernamePasswordAuthenticationToken(request.getEmail(),request.getPassword()));
     }
 
     @PostMapping("/logout")
-    public ResponseEntity<Void> logout(HttpServletRequest request){
+    public ResponseEntity<Void> logout(HttpServletRequest request) {
         String refreshToken = refreshTokenService.getRefreshTokenFromCookies(request);
-        if(refreshToken != null) {
+        if (refreshToken != null && !refreshToken.isEmpty()) {
             refreshTokenService.deleteByToken(refreshToken);
         }
         ResponseCookie jwtCookie = jwtService.getCleanJwtCookie();
         ResponseCookie refreshTokenCookie = refreshTokenService.getCleanRefreshTokenCookie();
         return ResponseEntity.ok()
-                .header(HttpHeaders.SET_COOKIE,jwtCookie.toString())
-                .header(HttpHeaders.SET_COOKIE,refreshTokenCookie.toString())
+                .header(HttpHeaders.SET_COOKIE, jwtCookie.toString())
+                .header(HttpHeaders.SET_COOKIE, refreshTokenCookie.toString())
                 .build();
-
     }
-
 }
